Migrated a lot of stuff from Portage:  - Streamlining of includes, library linking and other miscellaneous cleanup,    starting to prepare for compilation with g++ 4.3.  - BOOST_ROOT config variable  - ICU_ROOT config variable (started to do, but not complete yet)  - New Unit testing using CxxTest; moved legacy testing into tests/legacy.  - removed some obsolete programs: find_sentence_phrases, merge_ttables,    maxphrase.pl.  - Updated the whole LM library, with the refactored class hierarchy, and the    new Factory classes  - New LMDynMap class with number mapping and case mapping variants.  - Statistics of LM queries.  - Renamed lmtext2binlm arpalm2binlm, a better self-documenting name.  - Updated the tokenizer and detokenizer.  - New HMM aligner module, with all its variants, and a generic HMM toolkit to    support it.  (Uses dense vector/matrix representation, so not as generically    usable as it could be.)  - cat.sh: parallel word alignment training.  - Binary format for TTables and IBM1/2/HMM count files, needed to make cat.sh    reasonably efficient.  - eval_word_alignment calculates f-measure against a reference alignment  - More efficient gen-jpt-parallel.sh, in particular requiring a lot less RAM,    using the new merge_counts, which can also be used to combine any type of    count files.  - New file handling-unaligned-words.txt explains how NULL-aligned / unaligned    words work.  - Symmetrized training for IBM1 and 2 and HMM.  - linkPosteriors methods for all aligners, with the possibility of using them    for phrase extraction in gen_phrase_tables, via the PosteriorAligner.  - -prune1 switch to gen_phrase_tables and joint2cond_phrase_tables.  - Use tr1::unordered_map instead of the compiler specific hash_map.  - Refactor the different formats to represent word alignments into    word_align_io.{h,cc}, adding some new formats, and added word_align_tool to    manipulate and reformat word alingments.  - New truecasing modules supports learning the case mapping file, and    performating truecasing using canoe (default) or SRILM's disambig, for    users that have a valid SRILM licence.  Training of the language model is    not included, and requires that the user acquire appropriate third party    software, such as SRILM or IRSTLM.  - MagicStream support for lzma.  - Proper Casemapping for UTF-8 data using ICU; included as a utility class and    stand-alone program utf8_casemap, both supporting user-specified locales.  - utf8_filter validates and filters a stream, making sure it is valid UTF-8    and writing bad bytes sequences out in a way that they can be recognized and    dealt with.  - Utility class: BiVector, implementation for the abstraction of a vector with    positive and negative indices.  - Utility class template: tmp_val temporarily modifies a value.  - Various enhancements to diff-round.pl.  - Various improvements to run-parallel.sh.  - Removed strip-markup.pl, a script that doesn't do what it claims very well.  - Added canoe-escapes.pl to add or remove standard escapes needed by canoe.  - Utility class: WordClasses (needed by class based HMMs).

+++ src/tm/gen-jpt-parallel.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
-
 # gen-jpt-parallel.sh - generate a joint probability table in parallel
 #
 # PROGRAMMER: George Foster
 #
 # Technologies langagieres interactives / Interactive Language Technologies
-# Institut de technologie de l'information / Institute for Information Technology
+# Inst. de technologie de l'information / Institute for Information Technology
 # Conseil national de recherches Canada / National Research Council Canada
 # Copyright 2007, Sa Majeste la Reine du Chef du Canada /
 # Copyright 2007, Her Majesty in Right of Canada
@@ -67,7 +66,7 @@ is_int() {
 }
 
 NUM_JOBS=4
-OUTFILE=
+OUTFILE="-"
 RP_OPTS=
 VERBOSE=
 
@@ -93,6 +92,10 @@ if [ $# -lt 4 ]; then
 fi
 
 GPTARGS=($@)
+if [[ "$GPTARGS" =~ -v ]]; then
+   VERBOSE="-v"
+fi
+
 file1=${GPTARGS[$#-2]}
 file2=${GPTARGS[$#-1]}
 unset GPTARGS[$#-2]
@@ -126,6 +129,7 @@ if [ $DEBUG ]; then
     echo $WORKDIR >&2
     echo $NL >&2
     echo $SPLITLINES >&2
+    echo Verbose: $VERBOSE >&2
 fi
 
 # if [ ! $DEBUG ]; then
@@ -138,12 +142,10 @@ test -f $CMDS_FILE && \rm -f $CMDS_FILE
 for src in `ls $WORKDIR/L1.*`; do
    tgt=${src/\/L1./\/L2.}
    suff=${src##*/L1.}
-   echo "set -o pipefail; gen_phrase_tables -j ${GPTARGS[@]} $src $tgt | gzip > $WORKDIR/$suff.jpt.gz " >> $CMDS_FILE
+   echo "test ! -f $src || ((set -o pipefail; gen_phrase_tables -j ${GPTARGS[@]} $src $tgt | LC_ALL=C $SORT_DIR sort | gzip > $WORKDIR/$suff.jpt.gz) && mv $src $src.done)" >> $CMDS_FILE
 done
 
-if [ $DEBUG ]; then
-   cat $CMDS_FILE
-fi
+test $DEBUG && cat $CMDS_FILE
 
 if [ -n $VERBOSE ]; then
     echo "run-parallel.sh $VERBOSE $RP_OPTS $CMDS_FILE $NUM_JOBS" >&2
@@ -153,21 +155,20 @@ test $NOTREALLY ||
    eval "run-parallel.sh $VERBOSE $RP_OPTS $CMDS_FILE $NUM_JOBS"
 RC=$?
 if (( $RC != 0 )); then
-   echo "problems with run-parallel.sh (status=$RC) - quitting!" >&2
+   echo "problem calculating the joint frequencies (status=$RC) - quitting!" >&2
    exit $RC
 fi
 
-if [ $DEBUG ]; then
-   echo zcat $WORKDIR/*.jpt.gz \| joint2cond_phrase_tables -ij 
-fi
 
-if [ $NOTREALLY ]; then
-   :
-elif [ -n "$OUTFILE" ]; then
-   run-parallel.sh -nolocal -psub "-4" \
-       -e "zcat $WORKDIR/*.jpt.gz | joint2cond_phrase_tables -ij | gzip > $OUTFILE" 1
-else
-   zcat $WORKDIR/*.jpt.gz | joint2cond_phrase_tables -ij
+# Merging parts
+test $DEBUG && echo merge_counts -jpt $OUTFILE $WORKDIR/*.jpt.gz
+
+test $NOTREALLY ||
+   eval "merge_counts -jpt $OUTFILE $WORKDIR/*.jpt.gz"
+RC=$?
+if (( $RC != 0 )); then
+   echo "problems merging the joint frequencies (status=$RC) - quitting!" >&2
+   exit $RC
 fi
 
 rm -rf ${WORKDIR}
